Log everything to console
This way all logging will appear in the logs for whatever client is
running the server - which seems the neatest to me. For developers
running `./manage.py runserver`, logging will appear directly in the
terminal. For a server running `gunicorn`, the logs should be output
should be passed to `gunicorn`, for inclusion in its own logs.

The default log level in webapp/settings.py will be `ERROR`, so we don't
get too verbose logs, but it can be customised with the
`DJANGO_LOG_LEVEL` environment variable, and so we use that environment
variable to set the log level to `DEBUG` in local development.